requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
django-environ==0.4.3
django-money==0.11.2
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
gunicorn==19.7.1
idna==2.5
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
py-moneyed==0.7.0
PyJWT==1.5.0
pytz==2017.2
requests==2.17.3
six==1.10.0
urllib3==1.21.1
whitenoise==3.3.0
requests==2.17.3
